Gardens: Implement gardens list
